RDFIA/TME/Section2/2de/requirements.txt
torch>=1.10.2
torchvision>=0.11.3
pytorch-ignite>=0.4.8
pyyaml
omegaconf
hydra-core
wandb